confirm with password

+++ app/Livewire/Project/Shared/Danger.php
@@ -5,6 +5,8 @@ namespace App\Livewire\Project\Shared;
 use App\Jobs\DeleteResourceJob;
 use Livewire\Component;
 use Visus\Cuid2\Cuid2;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Auth;
 
 class Danger extends Component
 {
@@ -32,8 +34,13 @@ class Danger extends Component
         $this->environmentName = data_get($parameters, 'environment_name');
     }
 
-    public function delete()
+    public function delete($selectedActions, $password)
     {
+        if (!Hash::check($password, Auth::user()->password)) {
+            $this->addError('password', 'The provided password is incorrect.');
+            return;
+        }
+
         try {
             // $this->authorize('delete', $this->resource);
             $this->resource->delete();
